+++ src/com/example/stepnavi/filters/HighPassFilter.java
@@ -2,8 +2,8 @@ package com.example.stepnavi.filters;
 
 public class HighPassFilter {
 
-	private double current = 0.0f;
-	private double temp = 0.0f;
+	private float current = 0.0f;
+	private float temp = 0.0f;
 	
 	public void reset()
 	{
@@ -11,12 +11,12 @@ public class HighPassFilter {
 		temp = 	0.0f;
 	}
 	
-	public double getCurrentOutput()
+	public float getCurrentOutput()
 	{
 		return current;
 	}
 	
-	public double filter(double newValue, double magic)
+	public float filter(float newValue, float magic)
 	{
 		temp += (newValue - temp) / magic;
 		current = newValue - temp;
