+++ src/com/example/stepnavi/filters/LowPassFilter.java
@@ -2,19 +2,19 @@ package com.example.stepnavi.filters;
 
 public class LowPassFilter {
 
-	private double current = 0.0f;
+	private float current = 0.0f;
 	
 	public void reset()
 	{
 		current = 0.0f;
 	}
 	
-	public double getCurrentOutput()
+	public float getCurrentOutput()
 	{
 		return current;
 	}
 	
-	public double filter(double newValue, double magic)
+	public float filter(float newValue, float magic)
 	{
 		current += (newValue - current) / magic;	
 		return current;
